+++ types/src/jsMain/kotlin/screeps/arena/api/Source.kt
@@ -6,7 +6,7 @@ package screeps.arena.api
 /** An energy source object. Can be harvested by creeps with a {@link WORK} body part. */
 external open class Source : GameObject {
     /** Current amount of energy in the source. */
-    val energy: Double
+    val energy: Int
     /** The maximum amount of energy in the source. */
-    val energyCapacity: Double
+    val energyCapacity: Int
 }
