+++ types/src/jsMain/kotlin/screeps/arena/api/Store.kt
@@ -4,11 +4,11 @@
 package screeps.arena.api
 
 /** An object that class contain resources in its cargo. */
-external interface Store: Record<ResourceType, Double> {
+external interface Store: Record<ResourceType, Int> {
     /** Returns capacity of this store for the specified resource. */
-    fun getCapacity(resource: ResourceType = definedExternally): Double?
+    fun getCapacity(resource: ResourceType = definedExternally): Int?
     /** Returns free capacity for the store. */
-    fun getUsedCapacity(resource: ResourceType = definedExternally): Double?
+    fun getUsedCapacity(resource: ResourceType = definedExternally): Int?
     /** Returns the capacity used by the specified resource. */
-    fun getFreeCapacity(resource: ResourceType = definedExternally): Double?
+    fun getFreeCapacity(resource: ResourceType = definedExternally): Int?
 }
